Bug 916316 - Part 2: extract DateUtils.
