+++ test/src/org/mozilla/gecko/background/healthreport/test/TestHealthReportGenerator.java
@@ -11,6 +11,7 @@ import java.util.Iterator;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
+import org.mozilla.gecko.background.common.DateUtils;
 import org.mozilla.gecko.background.healthreport.HealthReportConstants;
 import org.mozilla.gecko.background.healthreport.HealthReportGenerator;
 import org.mozilla.gecko.background.healthreport.HealthReportStorage.Field;
@@ -250,7 +251,7 @@ public class TestHealthReportGenerator extends FakeProfileTestCase {
 
     long now = System.currentTimeMillis();
     JSONObject document = gen.generateDocument(0, 0, env1);
-    String today = HealthReportUtils.getDateString(now);
+    String today = DateUtils.getDateString(now);
 
     assertFalse(document.has("lastPingDate"));
     document = gen.generateDocument(0, HealthReportConstants.EARLIEST_LAST_PING, env1);
@@ -344,7 +345,7 @@ public class TestHealthReportGenerator extends FakeProfileTestCase {
 
     long now = System.currentTimeMillis();
     int day = storage.getDay(now);
-    final String todayString = HealthReportUtils.getDateString(now);
+    final String todayString = DateUtils.getDateString(now);
 
     int counter = storage.getField("org.mozilla.testm5", 1, "counter").getID();
     int discrete_int = storage.getField("org.mozilla.testm5", 1, "discrete_int").getID();
